fix: use int64_t seed and mt19937_64 for shuffle_after_epoch across all readers
- Change shuffle_after_epoch_seed_ member type from int32_t to int64_t
  in all loaders (IndexedFileLoader, WebdatasetLoader, FileLoader,
  FileLabelLoaderBase, NemoAsrLoader)
- Change TryGetArgument local seed variable from int32_t to int64_t in
  all constructors
- Update AddOptionalArg<int32_t> to AddOptionalArg<int64_t> in all
  reader op schemas (file, numpy, fits, coco, nemo_asr, tfrecord, mxnet,
  webdataset)
- Replace mt19937 seeded with (seed + epoch) with mt19937_64 seeded with
  seed + (epoch << 32) to avoid correlation between adjacent epoch seeds

+++ dali/operators/reader/coco_reader_op.cc
@@ -93,7 +93,7 @@ This readers produces the following outputs::
   .AddOptionalArg("shuffle_after_epoch",
       R"code(If set to True, the reader shuffles the entire  dataset after each epoch.)code",
       false)
-  .AddOptionalArg<int32_t>("shuffle_after_epoch_seed",
+  .AddOptionalArg<int64_t>("shuffle_after_epoch_seed",
       R"code(Random seed for the dataset shuffling performed after each epoch.
 
 If not provided, a fixed default seed is used, which results in the same shuffling
